Change ASN fields to uint32

+++ pkg/agent/bgp/gobgp/gobgp.go
@@ -165,7 +165,7 @@ func convertGoBGPPeerToPeerStatus(peer *gobgpapi.Peer) *bgp.PeerStatus {
 	}
 	if conf := peer.GetConf(); conf != nil {
 		peerStatus.Address = conf.GetNeighborAddress()
-		peerStatus.ASN = int32(conf.GetPeerAsn())
+		peerStatus.ASN = conf.GetPeerAsn()
 	}
 	if ebgpMultiHop := peer.GetEbgpMultihop(); ebgpMultiHop != nil {
 		peerStatus.MultihopTTL = int32(ebgpMultiHop.GetMultihopTtl())
@@ -255,7 +255,7 @@ func convertPeerConfigToGoBGPPeer(peerConfig bgp.PeerConfig) (*gobgpapi.Peer, er
 	peer := &gobgpapi.Peer{
 		Conf: &gobgpapi.PeerConf{
 			NeighborAddress: peerConfig.Address,
-			PeerAsn:         uint32(peerConfig.ASN),
+			PeerAsn:         peerConfig.ASN,
 			AuthPassword:    peerConfig.Password,
 		},
 		AfiSafis: []*gobgpapi.AfiSafi{
